added Lynn ann Austin template
I have made a rough template to what I think somone might want to put onto the about webpage. I have done the same template for Austin and Lynn and will wait for their feedback on what thry thought of it before continuing and asking the rest of the research group to complete.