attempt to make a collect information script using java
I was attempting to implement some code that would collect general information from each person in our lab group and display it generally on the front of the peoples page. As of right now I have been unable to fully implements the code, however, I think that if this script doesn't work now, I can get it to work after more time working in html.