My addition of Austin to the Website
This is the about me that I got back from Austin.  I was talking to her a bit thorughout the process and I was wondering what we could do to maybe make the students page a bit more interesting and I thought of adding a section that talks about why they are in the Lemas Lab group and what skills they think they're gaining/bringing to the lab. this would be agood way for them to talk about themselves more and still stay true to the research aspect of the website.

I think that this would be a nice thing to make the pages a bit more personal while keeping it still about the lab.

let me know if you want to branch out a bit more in the peoples pages or if you want to keep true to the general format.