made changes in people file type
all of the additonal people that were added to the website were added as a .txt file. I've noticed that Luran, magda, and dominick were saved as an md file. Becuase of this I've changed all file types to be an .md, hopefully this is the reason for the problems.