made modifications and sucessfully  fixed the issues with the people files
there was an issue in spacing that lead to a read error in the date they joined. additioanlly I caught a spelling error in the position section that lead to a misread so it didnt display the files. I tested the website on my local machine and it ran and displayed correctly.