made changes to travis file
removed jruby from running due to confliction with a gem in the gemfile.lock not installing and causing a build issue. additionally, changed the deployment from --deployment inline 13 to a separate call inline 14. this traivs file passed in my personal repo and I will now try merging to see if there is a build error in this repo.